release: Bump version to 1.0.2 with build system and header improvements
- Bump version to 1.0.2 in CMakeLists.txt, src/sonare.h, bindings/python/pyproject.toml, and package.json
- Refactor warning flags from global add_compile_options() to per-target SONARE_WARNING_FLAGS
  variable applied via target_compile_options() in src/, tests/, and tools/ to prevent
  warnings from propagating into third-party targets
- Add feature/cqt.h, feature/pitch.h, and feature/vqt.h to public includes in src/sonare.h
- Add POSITION_INDEPENDENT_CODE ON to shared library target in src/CMakeLists.txt
- Add -DCMAKE_BUILD_TYPE=Release to WASM build script in package.json
- Add release to .PHONY targets in Makefile
- Cast RAND_MAX to float in hpss_test.cpp to fix implicit conversion warning

+++ tests/effects/hpss_test.cpp
@@ -39,7 +39,8 @@ Audio create_percussive_audio(int sr = 22050, float duration = 0.5f, int n_click
     int start = c * interval;
     for (int i = 0; i < click_length && start + i < n_samples; ++i) {
       float envelope = 1.0f - static_cast<float>(i) / click_length;
-      samples[start + i] = envelope * (static_cast<float>(std::rand()) / RAND_MAX * 2.0f - 1.0f);
+      samples[start + i] =
+          envelope * (static_cast<float>(std::rand()) / static_cast<float>(RAND_MAX) * 2.0f - 1.0f);
     }
   }
